core: delegates: create DelegateHavingMixIn and use for Target
